requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
pyftpdlib==2.0.1
pytest==8.3.4
pytest-cov==6.0.0
gunicorn==23.0.0
werkzeug==3.1.3
